+++ R/ctl_pillar.R
@@ -114,6 +114,7 @@ rowidformat2 <- function(data, names, has_star) {
 
 #' Construct a custom pillar object
 #'
+#' @description
 #' `r lifecycle::badge("experimental")`
 #'
 #' `new_pillar()` is the low-level constructor for pillar objects.
@@ -172,9 +173,7 @@ format.pillar <- function(x, width = NULL, ...) {
     width <- sum(widths) - length(widths) + 1L
   }
 
-  out <- pillar_format_parts_2(x, width)
-
-  new_vertical(unlist(unname(out)))
+  new_vertical(pillar_format_parts_2(x, width))
 }
 
 #' @export