+++ R/type-sum.R
@@ -31,8 +31,15 @@ type_sum.factor <- function(x) {
 
 #' @export
 type_sum.default <- function(x) {
+  pillar_attr <- attr(x, "pillar")
+
+  label <- pillar_attr$label
+  if (!is.null(label)) {
+    return(I(label))
+  }
+
   if (is.object(x) || vctrs::vec_is(x)) {
-    return(vctrs::vec_ptype_abbr(x))
+    return(vec_ptype_abbr(x))
   }
 
   switch(typeof(x),
@@ -53,11 +60,11 @@ type_sum.default <- function(x) {
 
 # Registered in .onLoad()
 vec_ptype_abbr.pillar_empty_col <- function(x, ...) {
-  vctrs::vec_ptype_abbr(x[[1]])
+  vec_ptype_abbr(x[[1]])
 }
 
 #' @description
-#' `obj_sum()` also includes the size of the object if `vctrs::vec_is()`
+#' `obj_sum()` also includes the size of the object if [vctrs::vec_is()]
 #' is `TRUE`.
 #' It should always return a string (a character vector of length one).
 #'
@@ -104,7 +111,7 @@ size_sum <- function(x) {
 
 #' @export
 size_sum.default <- function(x) {
-  if (!vctrs::vec_is(x)) {
+  if (!vec_is(x)) {
     return("")
   }