+++ R/zzz.R
@@ -3,16 +3,45 @@
 #' @import rlang
 #' @import ellipsis
 #' @import lifecycle
+#' @importFrom vctrs data_frame
+#' @importFrom vctrs new_data_frame
+#' @importFrom vctrs obj_print_footer
+#' @importFrom vctrs obj_print_header
+#' @importFrom vctrs obj_print_data
+#' @importFrom vctrs s3_register
+#' @importFrom vctrs vec_arith
+#' @importFrom vctrs vec_arith.numeric
+#' @importFrom vctrs vec_arith_base
+#' @importFrom vctrs vec_c
+#' @importFrom vctrs vec_cast
+#' @importFrom vctrs vec_cast.double
+#' @importFrom vctrs vec_cast.integer
+#' @importFrom vctrs vec_data
+#' @importFrom vctrs vec_is
+#' @importFrom vctrs vec_math
+#' @importFrom vctrs vec_math_base
+#' @importFrom vctrs vec_proxy_compare
+#' @importFrom vctrs vec_proxy_order
+#' @importFrom vctrs vec_ptype_abbr
+#' @importFrom vctrs vec_ptype_full
+#' @importFrom vctrs vec_ptype2
+#' @importFrom vctrs vec_rbind
+#' @importFrom vctrs vec_restore
+#' @importFrom vctrs vec_size
 NULL
 
+# nolint start
 .onLoad <- function(libname, pkgname) {
+# nolint end
   # Can't use vctrs::s3_register() here with vctrs 0.1.0
   # https://github.com/r-lib/vctrs/pull/314
   register_s3_method("knitr", "knit_print", "pillar_squeezed_colonnade")
   register_s3_method("vctrs", "vec_ptype_abbr", "pillar_empty_col")
   register_s3_method("bit64", "pillar_shaft", "integer64", gen_pkg = "pillar")
   register_s3_method("survival", "pillar_shaft", "Surv", gen_pkg = "pillar")
+  register_s3_method("survival", "type_sum", "Surv", gen_pkg = "pillar")
   register_s3_method("survival", "pillar_shaft", "Surv2", gen_pkg = "pillar")
+  register_s3_method("survival", "type_sum", "Surv2", gen_pkg = "pillar")
 
   assign_crayon_styles()
 
@@ -23,7 +52,7 @@ NULL
   }
 
   # Necessary to re-parse environment variable
-  if (requireNamespace("debugme", quietly = TRUE)) {
+  if (Sys.getenv("DEBUGME") != "" && requireNamespace("debugme", quietly = TRUE)) {
     # activate_debugme()
     debugme::debugme()
     debug_info()