+++ tests/testthat/test-format_scientific.R
@@ -1,4 +1,4 @@
-format_scientific_bw <- function(x, ...) {
+split_decimal_bw <- function(x, ...) {
   ret <- pillar_shaft(x, ...)
   # Hack: Pretend decimal format requires 100 characters
   ret$dec <- set_width(ret$dec, 100)
@@ -8,23 +8,26 @@ format_scientific_bw <- function(x, ...) {
 }
 
 test_that("negative values displayed correct", {
-  f <- format_scientific_bw(-0.123)
+  f <- split_decimal_bw(-0.123)
   expect_equal(unname(format(f)), "-1.23e-1")
 })
 
 test_that("exponents correct in presence of NA", {
-  f <- format_scientific_bw(c(NA, 1e-5))
-  expect_equal(unname(format(f, justify = "right")), c("NA      ", " 1.00e-5"))
+  expect_snapshot({
+    format(split_decimal_bw(c(NA, 1e-5)), justify = "right")
+  })
 })
 
 test_that("corner cases", {
-  expect_equal(format_scientific_bw(numeric()), character())
+  expect_equal(split_decimal_bw(numeric()), character())
 })
 
 test_that("output test", {
-  expect_snapshot(pillar(add_special(10^c(-9, -6, 3, 9)), width = 10))
-  expect_snapshot(pillar(add_special((10^c(3, 9, 15, 22)) * c(-1, 1)), width = 10))
-  expect_snapshot(pillar(add_special(1.25 * 10^(-309:-319)), width = 10))
+  expect_snapshot({
+    pillar(add_special(10^c(-9, -6, 3, 9)), width = 10)
+    pillar(add_special((10^c(3, 9, 15, 22)) * c(-1, 1)), width = 10)
+    pillar(add_special(1.25 * 10^(-309:-319)), width = 10)
+  })
 })
 
 expect_scientific_width <- function(x) {
@@ -33,8 +36,8 @@ expect_scientific_width <- function(x) {
   }
 
   expect_equal(
-    get_formatted_width(format_scientific(!!x, 3)),
-    get_width(format_scientific(x, 3))
+    get_formatted_width(split_decimal(!!x, 3)),
+    get_width(split_decimal(x, 3))
   )
 }